Added taxonomy info for taxa with 16S sequences, modified how info is looked up when "host" is chosen
When "host" is chosen, the genus is first looked up in the "family" list, then it is looked up in "other"
If it is present in family but not in other, it is assumed to be a lichen photobiont.
If it is in other, that info is used.

There are more efficient ways to do this, but this is working for now.