Fixed error checking for sequences lacking CDS feature with a gene quality called <gene>
Script is supposed to print a warning when no CDS feature with gene quality called <gene>
is present in a sequence, but instead it was printing a duplicate of the CDS sequence from
the previous sequence. This is because the gene_seq variable was not being reinitialized
unless a new CDS sequence was found. Also, adding the empty gene_seq to gene gene list was
not generating an exception (instead there were exceptions when trying to print the empty
gene_seq), so I modified the code to explicitly check if a gene is present rather than
catching an exception