added code to label clades
This is an ugly hack because I can't figure out how to move the internode labels to
the right of the figure. Instead I am modifying the SVG after rendering. The way I have
done this is hacky and brittle. It would be far, far better to use a proper XML parser to
find the relavent nodes, but that is for a future version.