give singletons group names
This is necessary because of sequences that are duplicated in the DB
(because a single GB sequence represents multiple specimens with the
same haplotype). Previously they were handled correctly when they were
part of a larger group but if the sequence is a singleton, each
duplicate was labeled as unique and skipped when adding metadata to the
tree. Now they will all get assigned the same group name and will be
handled correctly.

This required some changes to the script to add the metadata to the
tree because it was relying on the presence of the work ‘Group” in the
group name (as opposed to UNIQUE) to decided wether to use the group
name or the accession number for the name on the tree. I changed this
to use the group size instead. This also allowed me to clean up the
code some, which should make it simpler and easier to maintain I do
need to test it still tho